Removed image metadata and made description optional
The image metadata for a post was used to have a top-level image for
each post. I'm not interested in enforcing that behavior or even using
it casually.

The description is an optional part of both RSS and Atom syndication. As
such, I don't believe it has any place being mandated by this blogging
tool.